Only download metadata in parallel
This commit builds on top of the parallel dependency resolution, but limits it to parallel download of metadata. It introduces
an API to determine if the network is effectively going to be hit, in which case edges are added to the queue of work to be
done in parallel. If not (locally available), everything is done serially. This is done because of important contention on
the global lock coordination service lock, which kills performance.